Refactor navbar and home page structure; remove unused components and update imports
