+++ src/app/globals.css
@@ -3,8 +3,6 @@
 @import "@fontsource/montserrat/700.css";
 @import "@fontsource/ysabeau/400.css";
 @import "@fontsource/ysabeau/600.css";
-@import "@fontsource/roboto/400.css";  
-@import "@fontsource/roboto/700.css";  
 
 :root {
   --background: #ffffff;
