+++ src/components/molecules/navbar.tsx
@@ -29,7 +29,7 @@ export default function Navbar() {
   }
 
   return (
-    <nav className="bg-black border border-white border-[0.1px] rounded-[40px] relative mx-auto max-w-[95%] lg:max-w-[90%] xl:max-w-[80%] mt-10">
+    <nav className="bg-black border-white border-[0.1px] rounded-[40px] relative mx-auto max-w-[95%] lg:max-w-[90%] xl:max-w-[80%] mt-10">
       <div className="w-full mx-auto px-2 sm:px-4 md:px-6 lg:px-8">
         <div className="flex justify-between h-16">
           <div className="flex items-center">
@@ -76,45 +76,13 @@ export default function Navbar() {
                 Interior Design
               </Link>
 
-              {/* Dropdown for Furniture */}
-              <div className="relative" ref={dropdownRef}>
-                <button
-                  className="flex items-center text-gray-300 hover:text-white px-2 lg:px-3 py-2 text-xs lg:text-sm font-medium"
-                  onClick={() => toggleDropdown("furniture")}
-                  style={{ fontFamily: "Montserrat" }}
-                >
-                  Furniture
-                  <ChevronDown className="ml-1 h-4 w-4" />
-                </button>
-
-                {activeDropdown === "furniture" && (
-                  <div className="absolute left-0 mt-2 w-32 md:w-36 bg-black border border-yellow-500 shadow-lg z-50">
-                    <div className="py-1">
-                      <Link
-                        href="/furniture/link-five"
-                        className="block px-4 py-2 text-sm text-gray-300 hover:bg-gray-800"
-                        style={{ fontFamily: "Montserrat" }}
-                      >
-                        Link Five
-                      </Link>
-                      <Link
-                        href="/furniture/link-six"
-                        className="block px-4 py-2 text-sm text-gray-300 hover:bg-gray-800"
-                        style={{ fontFamily: "Montserrat" }}
-                      >
-                        Link Six
-                      </Link>
-                      <Link
-                        href="/furniture/link-seven"
-                        className="block px-4 py-2 text-sm text-gray-300 hover:bg-gray-800"
-                        style={{ fontFamily: "Montserrat" }}
-                      >
-                        Link Seven
-                      </Link>
-                    </div>
-                  </div>
-                )}
-              </div>
+              <Link
+                href="/furniture"
+                className="text-gray-300 hover:text-white px-2 lg:px-3 py-2 text-xs lg:text-sm font-medium"
+                style={{ fontFamily: "Montserrat" }}
+              >
+                Furniture
+              </Link>
 
               <Link
                 href="/packages"
