+++ src/template/layout-context.tsx
@@ -1,14 +1,14 @@
 "use client";
 
-import Footer from "@/components/molecules/footer";
+import Navbar from "@/components/molecules/navbar";
 
 export function LayoutContent({ children }: { children: React.ReactNode }) {
     return (
       <div className="min-h-screen flex flex-col">
+        <Navbar />
         <main className="flex-grow w-full">
           {children}
         </main>
-        <Footer />
       </div>
     );
